enable preferences component to work with Oracle database backend and be backward compatible with HSQL involved using BIGINT because of the way that OJB uses setObject on its JDBC objects also involved converting ints to longs, shorts to ints
PR:
Obtained from:
Submitted by:	
Reviewed by:	
CVS: ----------------------------------------------------------------------
CVS: PR:
CVS:   If this change addresses a PR in the problem report tracking
CVS:   database, then enter the PR number(s) here.
CVS: Obtained from:
CVS:   If this change has been taken from another system, such as NCSA,
CVS:   then name the system in this line, otherwise delete it.
CVS: Submitted by:
CVS:   If this code has been contributed to Apache by someone else; i.e.,
CVS:   they sent us a patch or a new module, then include their name/email
CVS:   address here. If this is your work then delete this line.
CVS: Reviewed by:
CVS:   If we are doing pre-commit code reviews and someone else has
CVS:   reviewed your changes, include their name(s) here.
CVS:   If you have not had it reviewed then delete this line.


git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@186416 13f79535-47bb-0310-9956-ffa450edef68

+++ components/prefs/src/java/org/apache/jetspeed/prefs/om/Node.java
@@ -31,13 +31,13 @@ public interface Node extends Serializable, Cloneable
      * <p>Getter for the node id.</p>
      * @return The node id.
      */
-    int getNodeId();
+    long getNodeId();
 
     /**
      * <p>Setter for the node id.</p>
      * @param nodeId The node id.
      */
-    void setNodeId(int nodeId);
+    void setNodeId(long nodeId);
 
     /**
      * <p>Getter for the parent node id.</p>
@@ -45,13 +45,13 @@ public interface Node extends Serializable, Cloneable
      * is associated to a node.</p>
      * @return The parent node id.
      */
-    Integer getParentNodeId();
+    Long getParentNodeId();
 
     /**
      * <p>Setter for the parent node id.</p>
      * @param parentNodeId The parent node id.
      */
-    void setParentNodeId(Integer parentNodeId);
+    void setParentNodeId(Long parentNodeId);
 
     /**
      * <p>Getter for the node properties.</p>
@@ -97,7 +97,7 @@ public interface Node extends Serializable, Cloneable
      * </ul>
      * @return The node type.
      */
-    short getNodeType();
+    int getNodeType();
 
     /**
      * <p>Setter for the node type.</p>
@@ -107,7 +107,7 @@ public interface Node extends Serializable, Cloneable
      * </ul>
      * @param nodeType The node type.
      */
-    void setNodeType(short nodeType);
+    void setNodeType(int nodeType);
 
     /**
      * <p>Getter for the full path.</p>
